bot/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
requests==2.31.0
PyJWT>=2.0.0
dateparser==1.2.1